airflow/requirements.txt
boto3
botocore
datasets
DateTime==5.5
Deprecated==1.2.14
huggingface-hub
mysql-connector==2.2.9
mysql-connector-python
numpy
pandas
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-dotenv
pytz==2024.2
PyYAML
requests
requests-toolbelt
s3transfer
setuptools
apache-airflow==2.10.2
unstructured-ingest==0.0.23
unstructured-ingest[s3]
unstructured-ingest[remote]
pymupdf==1.24.11
pymupdf4llm==0.0.17
pillow==10.4.0